Strip down to single block widget
